+++ src/repository/injection.go
@@ -45,6 +45,7 @@ func GetInjectionByID(db *gorm.DB, id int) (*database.FaultInjection, error) {
 	var injection database.FaultInjection
 	if err := db.
 		Preload("Task").
+		Preload("Task.Trace").
 		Preload("Benchmark.Container").
 		Preload("Pedestal.Container").
 		Where("id = ?", id).First(&injection).Error; err != nil {